Major Fixes
this commit contains major fixes to the website. file checking was not done on every file in the commit since the changed files were too many and this commit holds no major changes feature wise but merely some cleanup and making the website actually work. some of the fixes include but are not limited to:
- changed Class field names to become camelCase
- right now a user can register, login, edit his info and all the basic user management operations.
- a logged in user can create multiple bank accounts and can view them through the BankAccounts URL.
- a user can only view his bank accounts and no one else's.